Build the useSort hook
