+++ src/features/device/components/device-table-columns.tsx
@@ -2,61 +2,25 @@
 
 import { cn } from "@/lib/utils";
 
-import { useState } from "react";
-
 import { ChevronUpIcon } from "lucide-react";
-
-import { usePathname, useRouter, useSearchParams } from "next/navigation";
 
 import { getDeviceTableColumns } from "@/features/device/lib/utils";
 
 import type { SortDirection } from "@/features/device/lib/definitions";
 
-import { SORT_BY_PARAM_ID, SORT_DIRECTION_PARAM_ID } from "@/features/device/lib/constants";
+import useSort from "@/features/device/hooks/use-sort";
 
 import { TableHead } from "@/components/ui/table";
 
 type DeviceTableColumnsProps = ReturnType<typeof getDeviceTableColumns>;
 
 export default function DeviceTableColumns({ columns }: { columns: DeviceTableColumnsProps }) {
-  const pathname = usePathname();
-
-  const { replace } = useRouter();
-
-  const searchParams = useSearchParams();
-
-  const getInitialState = () => {
-    const sortBy = searchParams.get(SORT_BY_PARAM_ID) || "";
-    const sortDirection = searchParams.get(SORT_DIRECTION_PARAM_ID) || "asc";
-
-    if (!sortBy && !sortDirection) return null;
-
-    return {
-      name: sortBy,
-      direction: sortDirection as SortDirection,
-    };
-  };
-
-  const [sortBy, setSortBy] = useState<{ name: string; direction: SortDirection } | null>(getInitialState());
+  const { sort, handleSort } = useSort();
 
   return columns.map((tableColumn) => {
     const isActionsColumn = tableColumn.toLowerCase() === "actions";
 
-    const handleSort = (value: string) => {
-      const newSortByDirection = sortBy?.name === value.toLowerCase() && sortBy?.direction === "asc" ? "desc" : "asc";
-
-      setSortBy({ name: value.toLowerCase(), direction: newSortByDirection });
-
-      const params = new URLSearchParams(searchParams);
-
-      params.set(SORT_BY_PARAM_ID, value.toLowerCase());
-
-      params.set(SORT_DIRECTION_PARAM_ID, newSortByDirection);
-
-      replace(`${pathname}?${params.toString()}`);
-    };
-
-    const isSortedColumn = tableColumn.toLowerCase() === sortBy?.name;
+    const isSortedColumn = tableColumn.toLowerCase() === sort?.name;
 
     return (
       <TableHead
@@ -66,7 +30,7 @@ export default function DeviceTableColumns({ columns }: { columns: DeviceTableCo
       >
         <div className="flex items-center gap-1">
           <span className={isActionsColumn ? "sr-only" : undefined}>{tableColumn}</span>
-          {!isActionsColumn && isSortedColumn && <SortIcon direction={sortBy.direction} />}
+          {!isActionsColumn && isSortedColumn && <SortIcon direction={sort.direction} />}
         </div>
       </TableHead>
     );
